RGIT/backend/requirements.txt
fastapi<0.100.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic<2
librosa==0.10.2
numpy
opencv-python-headless>=4.9.0
ultralytics>=8.2.0
ffmpeg-python==0.2.0
python-dotenv==1.0.1
aiofiles==24.1.0
speechrecognition==3.10.1
textblob==0.18.0.post0
groq>=1.0.0
httpx>=0.27.0
yt-dlp>=2024.1.0
xhtml2pdf>=0.2.16
scikit-learn>=1.4.0
xgboost>=2.0.0
ollama>=0.3.0
